Move workflow related test
